feat(tags): extract tags package from utils. Update all references
