+++ src/pkg/tags/tags_test.go
@@ -1,30 +1,30 @@
-package utils_test
+package tags_test
 
 import (
 	"reflect"
 	"testing"
 
-	"github.com/defenseunicorns/go-oscal/src/pkg/utils"
+	"github.com/defenseunicorns/go-oscal/src/pkg/tags"
 )
 
 func TestTagsUtils(t *testing.T) {
 	t.Parallel()
 
-	t.Run("splits comma separated tags", func(t *testing.T) {
+	t.Run("splits comma separated inputTags", func(t *testing.T) {
 		t.Parallel()
-		tags := "json,yaml"
+		inputTags := "json,yaml"
 		expected := []string{"json", "yaml"}
-		actual, _ := utils.FormatTags(tags)
+		actual, _ := tags.FormatTags(inputTags)
 		if !reflect.DeepEqual(actual, expected) {
 			t.Error("expected", expected, "got", actual)
 		}
 	})
 
-	t.Run("defaults to json if no tags are provided", func(t *testing.T) {
+	t.Run("defaults to json if no inputTags are provided", func(t *testing.T) {
 		t.Parallel()
-		tags := ""
+		inputTags := ""
 		expected := []string{"json"}
-		actual, _ := utils.FormatTags(tags)
+		actual, _ := tags.FormatTags(inputTags)
 		if !reflect.DeepEqual(actual, expected) {
 			t.Error("expected", expected, "got", actual)
 		}
@@ -32,9 +32,9 @@ func TestTagsUtils(t *testing.T) {
 
 	t.Run("returns an error if an invalid tag is provided", func(t *testing.T) {
 		t.Parallel()
-		tags := "json,xml"
+		inputTags := "json,xml"
 		expected := "invalid tag: xml\n"
-		_, err := utils.FormatTags(tags)
+		_, err := tags.FormatTags(inputTags)
 		if err.Error() != expected {
 			t.Error("expected", expected, "got", err)
 		}
